Added initial random generation of NPC
There are two kinds of NPCs to be generated: Customers and FreeRoamers.

I added a separate class for Customers and FreeRoamers as well a ShopManager that handles the random generation of Orders for each Customers.

The problem lies on how to make the NPC appear in the game.

WELP, lets check the code together bwaha